control_server/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.10.0
pytest==7.2.0
pytest-flask==1.2.0
python-decouple==3.6
pymongo==4.3.2
munch==2.5.0
requests==2.28.1
Brotli==1.0.9
fastcrc==0.2.1